Add custom sort algorithms for by item type and by item quality
Will sort the entire bank, including reagent bank and it will move any
reagents in the regular bank to any empty slots in the reagent bank.

Special bags (herbalism/inscription/etc) are supported.

Hearthstones are sorted first

Also permits as an option changing the sort algorithm to the Blizzard
one (retail) and to SortBags (classic, if available)