New corners' widget priorities options
Every corner can have multiple widgets assigned, and the first priority
widget that displays will be used. This allows setting pawn and then
item level to the same corner, so that the item level only shows when
the pawn icon doesn't.

Added API for other addons to add their own widgets to the corners